add fronted and config. bug fixes

+++ src/main/java/com/example/college_management/service/DepartmentService.java
@@ -37,6 +37,9 @@ public class DepartmentService {
     }
 
     public void deleteDepartment(Long id) {
+        if (!departmentRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Department Not Found!");
+        }
         departmentRepository.deleteById(id);
     }
 }
